requirements-test.txt
pytest>=6,<7
pytest-sqlalchemy>=0.2,<0.3
testfixtures>=6,<7
